requirements.txt
# AI Study Assistant Multi-Agent System Requirements

# Core Framework
crewai>=0.1.0
crewai-tools>=0.0.6

# LLM Integration and Language Models
openai>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.1

# Environment Configuration
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
pydantic>=2.0.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
